Add worker_nodes_subnet to its new location
An upstream patch[0] changed the group where the worker_nodes_subnet
resides. That obviously breaks installation as soon as the change is
merged and a new delorean package gets released

This patch keeps it in both places so that we can merge this right away
and be ready for when the upstream patch makes it into the packaging.

[0] https://review.openstack.org/#/c/450113/

Change-Id: Ie0997b000f400777b3e66f7e55d8ecb2f9dd6c88
Signed-off-by: Antoni Segura Puimedon <celebdor@gmail.com>